+++ packages/uploadthing/src/internal/types.ts
@@ -1,10 +1,7 @@
-import type * as Config from "effect/Config";
-import type * as LogLevel from "effect/LogLevel";
 import type { Schema } from "effect/Schema";
 
 import type {
   ErrorMessage,
-  FetchEsque,
   FileRouterInputConfig,
   Json,
   JsonObject,
@@ -14,7 +11,6 @@ import type {
   UploadThingError,
 } from "@uploadthing/shared";
 
-import type { LogFormat } from "./logger";
 import type { JsonParser } from "./parser";
 import type {
   FileUploadDataWithCustomId,
@@ -47,11 +43,6 @@ export type UnsetMarker = typeof unsetMarker;
 export type ValidMiddlewareObject = {
   [UTFiles]?: Partial<FileUploadDataWithCustomId>[];
   [key: string]: unknown;
-};
-
-type ResolverOptions<TMetadata> = {
-  metadata: TMetadata;
-  file: UploadedFileData;
 };
 
 /**
@@ -87,9 +78,10 @@ type MiddlewareFn<
   },
 ) => MaybePromise<TOutput>;
 
-type UploadCompleteFn<TMetadata, TOutput extends JsonObject | void> = (
-  opts: ResolverOptions<TMetadata>,
-) => MaybePromise<TOutput>;
+type UploadCompleteFn<TMetadata, TOutput extends JsonObject | void> = (opts: {
+  metadata: TMetadata;
+  file: UploadedFileData;
+}) => MaybePromise<TOutput>;
 
 type UploadErrorFn = (input: {
   error: UploadThingError;
@@ -160,13 +152,13 @@ export interface UploadBuilder<TParams extends AnyParams> {
   }>;
 }
 
-export type BuiltUploaderTypes = {
+export type AnyBuiltUploaderTypes = {
   input: any;
   output: any;
   errorShape: any;
 };
 
-export interface Uploader<TTypes extends BuiltUploaderTypes> {
+export interface Uploader<TTypes extends AnyBuiltUploaderTypes> {
   $types: TTypes;
   routerConfig: FileRouterInputConfig;
   routeOptions: RouteOptions;
@@ -176,66 +168,7 @@ export interface Uploader<TTypes extends BuiltUploaderTypes> {
   errorFormatter: (err: UploadThingError) => any;
   onUploadComplete: UploadCompleteFn<any, any>;
 }
-export type AnyUploader = Uploader<BuiltUploaderTypes>;
-
-export type FileRouter = Record<string, AnyUploader>;
-
-export type RouteHandlerConfig = {
-  logLevel?: LogLevel.Literal;
-  /**
-   * What format log entries should be in
-   * @default "pretty" in development, else "json"
-   * @see https://effect.website/docs/guides/observability/logging#built-in-loggers
-   */
-  logFormat?: Config.Config.Success<typeof LogFormat>;
-  callbackUrl?: string;
-  token?: string;
-  /**
-   * Used to determine whether to run dev hook or not
-   * @default `env.NODE_ENV === "development" || env.NODE_ENV === "dev"`
-   */
-  isDev?: boolean;
-  /**
-   * Used to override the fetch implementation
-   * @default `globalThis.fetch`
-   */
-  fetch?: FetchEsque;
-  /**
-   * Set how UploadThing should handle the daemon promise before returning a response to the client.
-   * You can also provide a synchronous function that will be called before returning a response to
-   * the client. This can be useful for things like:
-   * -  [`@vercel/functions.waitUntil`](https://vercel.com/docs/functions/functions-api-reference#waituntil)
-   * - [`next/after`](https://nextjs.org/blog/next-15-rc#executing-code-after-a-response-with-nextafter-experimental)
-   * - or equivalent function from your serverless infrastructure provider that allows asynchronous streaming
-   * If deployed on a stateful server, you most likely want "void" to run the daemon in the background.
-   * @remarks - `"await"` is not allowed in development environments
-   * @default isDev === true ? "void" : "await"
-   */
-  handleDaemonPromise?:
-    | "void"
-    | "await"
-    | ((promise: Promise<unknown>) => void);
-  /**
-   * URL override for the ingest server
-   */
-  ingestUrl?: string;
-};
-
-export type RouteHandlerOptions<TRouter extends FileRouter> = {
-  router: TRouter;
-  config?: RouteHandlerConfig;
-};
-
-export type inferEndpointInput<TUploader extends AnyUploader> =
-  TUploader["$types"]["input"];
-
-export type inferEndpointOutput<TUploader extends AnyUploader> =
-  TUploader["$types"]["output"] extends UnsetMarker | void | undefined
-    ? null
-    : TUploader["$types"]["output"];
-
-export type inferErrorShape<TRouter extends FileRouter> =
-  TRouter[keyof TRouter]["$types"]["errorShape"];
+export type AnyUploader = Uploader<AnyBuiltUploaderTypes>;
 
 /**
  * Map actionType to the required payload for that action
@@ -243,7 +176,7 @@ export type inferErrorShape<TRouter extends FileRouter> =
  */
 export type UTEvents = {
   upload: {
-    in: typeof UploadActionPayload.Type;
+    in: UploadActionPayload;
     out: ReadonlyArray<NewPresignedUrl>;
   };
 };
