+++ packages/uploadthing/src/next-legacy.ts
@@ -5,9 +5,9 @@ import type { Json } from "@uploadthing/shared";
 
 import { makeAdapterHandler } from "./internal/handler";
 import { toWebRequest } from "./internal/to-web-request";
-import type { FileRouter, RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
+import type { FileRouter, RouteHandlerOptions } from "./types";
 
 export { UTFiles } from "./internal/types";
 export type { FileRouter };
